+++ src/utils/offchainState.ts
@@ -5,6 +5,7 @@ import { createLogger } from '@subsquid/logger'
 import assert from 'assert'
 import { uniqueId } from './crypto'
 import { defaultNotificationPreferences } from './notification/helpers'
+import { NextEntityId } from '../model'
 
 const DEFAULT_EXPORT_PATH = path.resolve(__dirname, '../../db/export/export.json')
 
@@ -91,6 +92,13 @@ function migrateExportDataToV310(data: ExportedData): ExportedData {
 export class OffchainState {
   private logger = createLogger('offchainState')
   private _isImported = false
+
+  private globalCountersMigration = {
+    // destination version : [global counters names]
+    '3.0.1': ['Account'],
+    '3.0.2': ['Account'],
+    '3.1.0': ['Account'],
+  }
 
   private migrations: Migrations = {
     '3.1.0': migrateExportDataToV310,
@@ -264,5 +272,25 @@ export class OffchainState {
     const { blockNumber }: ExportedState = JSON.parse(fs.readFileSync(exportFilePath, 'utf-8'))
     this.logger.info(`Last export block number established: ${blockNumber}`)
     return blockNumber
+  }
+
+  private async migrateCounters(exportedVersion: string, em: EntityManager): Promise<void> {
+    const migrationData = Object.entries(this.globalCountersMigration).sort(
+      ([a], [b]) => this.versionToNumber(a) - this.versionToNumber(b)
+    ) // sort in increasing order
+
+    for (const [version, counters] of migrationData) {
+      if (this.versionToNumber(exportedVersion) < this.versionToNumber(version)) {
+        this.logger.info(`Migrating global counters to version ${version}`)
+        for (const entityName of counters) {
+          // build query that gets the entityName with the highest id
+          const rowNumber = await em.query(`SELECT COUNT(*) FROM ${entityName}`)
+          const latestId = parseInt(rowNumber[0].count)
+
+          this.logger.info(`Setting next id for ${entityName} to ${latestId + 1}`)
+          await em.save(new NextEntityId({ entityName, nextId: latestId + 1 }))
+        }
+      }
+    }
   }
 }